Abstract release methods
